Improve preDownloadSleep logic
